add logout button in side nav bar
