+++ src/components/Settings/Settings.jsx
@@ -1,6 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import { Routes, Route, NavLink } from 'react-router-dom';
 import { doc, getDoc } from 'firebase/firestore';
+import { signOut } from 'firebase/auth';
 
 import ExtensionSetting from './components/Pages/ExtensionSetting';
 import FeedbackSetting from './components/Pages/FeedbackSetting';
@@ -22,7 +23,7 @@ import { db } from '../../utils/FirebaseConfig';
 import { useAuth } from '../../contexts/AuthContextProvider';
 
 const Settings = () => {
-	const { user } = useAuth();
+	const { user, auth } = useAuth();
 	const [device, setDevice] = useState('');
 	const [deviceSpecification, setDeviceSpecification] = useState('');
 	const [hasDeviceSpecification, setHasDeviceSpecification] = useState(false);
@@ -87,6 +88,15 @@ const Settings = () => {
 			icon: InformationCircleIcon,
 		},
 	];
+
+	const logout = async () => {
+		try {
+			await signOut(auth);
+		} catch (error) {
+			return;
+		}
+	};
+
 	return (
 		<div className="grid grid-cols-[minmax(12rem,16rem),1fr] gap-4">
 			<div className="mi z-10 flex flex-col gap-4 bg-slate-100 drop-shadow-2xl">
@@ -107,8 +117,15 @@ const Settings = () => {
 						</div>
 					</NavLink>
 				))}
-			</div>
 
+				{user && (
+					<div className="flex h-full flex-col justify-end px-4 pb-8">
+						<button className="mt-10 h-10 w-full rounded-lg bg-slate-500 text-white" onClick={logout}>
+							Uitloggen
+						</button>
+					</div>
+				)}
+			</div>
 			<div className="">
 				<Routes>
 					<Route path="feedback" element={<FeedbackSetting />} />
